Moved GC Statistics into its own file
GC Statistics module has been moved to its own file.
